Fix telegram_id range issue in User model

+++ giveaway/models.py
@@ -30,7 +30,7 @@ class User(AbstractBaseUser, PermissionsMixin):
     last_name = models.CharField(max_length=100, null=True)
     user_name = models.CharField(max_length=100, unique=True, null=True, blank=True)
     phone_number = models.CharField(max_length=15, unique=True, null=True, blank=True)
-    telegram_id = models.IntegerField(unique=True, null=True, blank=True)
+    telegram_id = models.BigIntegerField(unique=True, null=True, blank=True)
     date_joined = models.DateTimeField(auto_now_add=True)
     language = models.CharField(max_length=100, default="english")
